refactor(SocketClient): 重构了TCP连接模块
